Decouple Application interface, TO BE CONTINUED
