+++ eagle-core/eagle-app/eagle-app-base/src/test/java/org/apache/eagle/app/TestApplicationTestSuite.java
@@ -26,7 +26,6 @@ import org.apache.eagle.metadata.model.ApplicationEntity;
 import org.apache.eagle.metadata.model.SiteEntity;
 import org.apache.eagle.metadata.resource.SiteResource;
 import org.junit.Assert;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -76,8 +75,8 @@ public class TestApplicationTestSuite {
         simulator.submit("TEST_APPLICATION");
     }
 
-    @Test
-    public void testApplicationQuickRunWithAppProvider(){
-        simulator.submit(TestApplicationImpl.Provider.class);
-    }
+//    @Test
+//    public void testApplicationQuickRunWithAppProvider(){
+//        simulator.submit(TestStormAppImpl.Provider.class);
+//    }
 }
