+++ eagle-core/eagle-app/eagle-app-base/src/test/resources/TestApplicationMetadata.xml
@@ -20,7 +20,7 @@
     <type>TEST_APPLICATION</type>
     <name>Test Monitoring Application</name>
     <version>0.5.0-incubating</version>
-    <appClass>org.apache.eagle.app.TestApplicationImpl</appClass>
+    <appClass>org.apache.eagle.app.TestStormAppImpl</appClass>
     <viewPath>/apps/example</viewPath>
     <configuration>
         <property>
