+++ eagle-examples/eagle-app-example/src/main/resources/META-INF/apps/example/metadata.xml
@@ -20,7 +20,7 @@
     <type>EXAMPLE_APPLICATION</type>
     <name>Example Monitoring Application</name>
     <version>0.5.0-incubating</version>
-    <appClass>org.apache.eagle.app.example.ExampleApplication</appClass>
+    <appClass>org.apache.eagle.app.example.ExampleStormApp</appClass>
     <viewPath>/apps/example</viewPath>
     <configuration>
         <property>
